Stream media downloads via signed routes instead of dehydrating through Livewire
Livewire's SupportFileDownloads intercepts every BinaryFileResponse /
StreamedResponse returned from a component, captures the body inside
ob_start() and base64-encodes it into the JSON response payload. For an
on-the-fly zip of attachments this triples memory usage on top of the
buffer itself - production hit "Allowed memory size of 134217728 bytes
exhausted" five different ways on /projects/1875 alone, all rooted in
WithFileUploads::downloadCollection.

Split WithFileUploads into two focused traits:

- SupportsFileUploads keeps the actual upload concern (Livewire's
  WithFileUploads base, prepareForMediaLibrary, removeFileUpload,
  saveFileUploadsToMediaLibrary).
- SupportsFileDownloads owns single-file and collection downloads, both
  of which now generate a signed URL and hand control back to the
  browser via $this->redirect(). No bytes ever pass through Livewire's
  payload.

WithFileUploads itself becomes a thin trait that composes both, so the
ten existing consumers keep working unchanged.

For the on-the-fly zip case, Livewire pre-validates the action data,
encrypts {data, name} into a 5-minute signed URL to a new route
/media-collection-download/{token}, and redirects. The route streams
the zip via response()->streamDownload() and ZipStream-PHP's
getZipStream() — chunked output to the browser, never buffered.

The four components that hand-rolled their own download helpers
(MediaList::downloadMedia, PurchaseInvoiceList::downloadMedia,
EditMail::downloadAttachment, Mail::download) drop those duplicates and
either pull SupportsFileDownloads in directly (MediaList, Mail) or get
it via WithFileUploads composition (EditMail, PurchaseInvoiceList).
The edit-mail blade switches its $wire.downloadAttachment(file.id)
call to $wire.download(file.id).

Tests assert the download / downloadCollection effects are now Livewire
redirects instead of SupportFileDownloads-injected 'download' effects.

BREAKING: $wire.downloadMedia / $wire.downloadAttachment no longer
exist on MediaList / EditMail — consumer blades that called them must
use $wire.download instead.
